[modify] 스토리 데이터 parcelable 수정

+++ app/src/main/java/org/personal/coupleapp/data/StoryData.kt
@@ -4,13 +4,13 @@ import android.os.Parcel
 import android.os.Parcelable
 
 @Suppress("UNREACHABLE_CODE")
-data class StoryData(val id: Int?, val couple_id: Int, val title: String, val description: String, val date: Int, var photo_path: ArrayList<Any>) : Parcelable {
+data class StoryData(val id: Int?, val couple_id: Int, val title: String, val description: String, val date: Long, var photo_path: ArrayList<Any>) : Parcelable {
     constructor(parcel: Parcel) : this(
         parcel.readValue(Int::class.java.classLoader) as? Int,
         parcel.readInt(),
         parcel.readString()!!,
         parcel.readString()!!,
-        parcel.readInt(),
+        parcel.readLong(),
         parcel.readArrayList(ArrayList::class.java.classLoader) as ArrayList<Any>
     )
 
@@ -19,7 +19,7 @@ data class StoryData(val id: Int?, val couple_id: Int, val title: String, val de
         parcel.writeInt(couple_id)
         parcel.writeString(title)
         parcel.writeString(description)
-        parcel.writeInt(date)
+        parcel.writeLong(date)
         parcel.writeList(photo_path as List<*>)
     }
